backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
orjson>=3.10
openai>=1.40
python-dotenv>=1.0
httpx>=0.27
